Bevy 0.16 compatibility and updates (#36)
* before removing gizmos

* after removing most of gizmo

* no red lines

* gets to first update

will crash as soon as egui Tries to render a font

* Include error in panic

* rename gizmo camera to ui camera

you have two add gizmo camera systems, one of witch actually adds the UiCamera component so I renamed it accordingly

* Fixed Egui crashing

turns out you need to run all the Egui systems in the EguiPrimaryContextPass schedul

* make add componets popup stay

the popup would close the next frame if you don't call keep open.
will have to replace egui impl since later it uses lots of deprecated functions

* chain ui systems

you have two copies of the same system so i assume they need to run in order

* fixed crash when adding MainCamera

something to do with how you where cloning reflected components.
I changed from deprecated clone_value() to relfect_clone().
reflect_clone return a Result so i log that if it happens, the key difference is reflect_clone keeps the underlying type intact allowing for clone to be lossless

* added /assets to gitignore

when you run the example it generates as assets folder and we don't want that to be part of the repo

* remove fix_ctx from example

this was here to test and i have since fixed the crash

* re-added gizmos

very buggy but need to fix other bugs first

* revert EditableMaterial to use handle

When I was porting I replaced all use of Handle<_> with the updated Component version, this was incorrect for this case since it truly is a handle and not the component

* fixed the all pink

this probably overkill/feature creep
Bevys default tonemap breaks when default-features = false, so I have added the ability to register bevy components to show up in the editor so I can change the tonemap to None and make everything not pink again

back to fixing gizmos, as far as I can tell its just rotation/drag to go.
but translation feels off because its not in sync with the cursor, seems to move about double what you drag

* added ability to show all entities in editor

this will help debug the editor its self, and will allow for create authors to hide entities but uses still be able to selectively see them

* show all components in unfiltered mode

show all reflect components when in unfiltered mode to allow for easy inspecting of components

* free rotate gizmo finally works

* fixed snap rotate

* added single axis rotate back

this should be done on a global level but im too tired for that now

* bump back to 0.16

* change transform gizmo to use observes

* added locked gizmo lines

* added CTRL makes camera move with gizmo

* i broke the history again

* :recycle: Fixes deprecated get_single

* :recycle: Uses child_of.parent

* :recycle: Uses single_mut

* :recycle: Uses modern equi APIs

Leaves one instead of keep_popup_open since that is not a straightforward API change

* :recycle: Uses EventWriter.write

* :recycle: Creates weak handle from uuid

* :recycle: Despawns recursively by default

* :construction: Uses Reflect Clone

* :recycle: Inserts ChildOf

* yellow line hunt

removed most yellow lines
didn't do any that might change logic
didn't do any that are related to format! args

* converted all Vec<String> to Vec<Cow>

* implement safe entity despawning to prevent crashes

Collect gizmo target entities and despawn them safely in selection cleanup

* Despawn gizmo on change

* Check for mouse over egui before selecting

From UserInput we have mouse_over_egui, so when we go to select an entity, check if we are not over egui

* Axis line Transform -> GlobalTransform (#33)

Also cleaned up old axis line rendering

* Transform gizmo cleanup (#34)

* begin cleanup

* Force children to move globally with transform gizmo

* Planes now respect global

---------

Co-authored-by: PhaestusFox <phaestusfox@gmail.com>
Co-authored-by: PhaestusFox <75610846+PhaestusFox@users.noreply.github.com>
Co-authored-by: Eric Kwoka <43540491+ekwoka@users.noreply.github.com>