make transform gizmo only use primary InputButton
Technically any two button clicks on the gizmo gets it in an invalid state since two sets of Drag events are sent and are in different axis Right mouse is just not as obvious because the camera is moving and breaks the raycast
Ps. Is there any reason we are using UserInput instead of bevy's input system?
feels like redundant work to maintain and prone to missmatch logic It also makes the editor less versatile since things like picking have a abstracted concept of "Primary" and "Secondary" button so it works independently of user control surface. eg touch, pen, mouse