added reduced format
the reduced format will use the type name only if it can.
if the type name is already taken it will make one step up the module path and use that, repeat till available name is found.
at the end of the scene file it adds a list of all name mappings.
{short} -> {full}

I could use the type_regustry to lookup if a name is_ambiguse, this just means two registered types share a name, then use the short name if not of the full name if, but I think for compatibility reasons a scene should keep note of the full name otherwise if you add a dependency that has a conflicting type there is no way of knowing what the original was. it also means that we don't have to fall back to full paths to resolve conflicts we can just step one module up