fixed crash is side panel is to large (#85)
* fixed crash

The side panel can be so wide we cant fit the bottom panel in the remaining space.
Hardcoded to 270px min since this is enough for now.
need a way to calculate min space needed so it doesn't start crashing again if we change contents of bottom bar.

* change from hardcoded scale

this is unrelated to the crash but change from scaling viewport 1.5 to scaling by ctx.pixels_per_point()

---------

Co-authored-by: Blake Darrow <71980368+BlakeDarrow@users.noreply.github.com>