Fix objects jitter when applying transformations. (#95)
When we apply transformations to the objects using mouse drag, the
affected object seem to jitter because transformation is immediately
applied on `observer` call. Which not necessary happens on the right
time in call schedule.

By extracting transform application to the separate function that runs
on update jittering issue is fixed.

Co-authored-by: Nikita V <nikita.vakhtsin@ergo.lt>