Fix gizmo to cursor snap (#97)
* Fix gizmo to cursor snap

Fixes an unpleasant snap of the object to the cursor when we start
position change.

* Fix entity deselect on drag end.

Looks like right now in bevy picking this sequence of
events:
1. `Pointer<Press>`
2. `Pointer<DragStart>`
...
3. `Pointer<DragEnd>`

Will provide a valid `Pointer<Click>` event. Which in turn for us
triggered unselect of selected object.

Validating only `original_event_target` instead of all pierced objects
gives this function ability to only know about click on the gizmo, which
will not trigger unselect.

---------

Co-authored-by: Nikita V <nikita.vakhtsin@ergo.lt>